+++ VisualEffects.cpp
@@ -5,7 +5,7 @@ void explosionEffect::DrawExprosion()
     angle += 0.02f;
     progress += oneFlame;
     size += 20 * oneFlame;
-    SetDrawBlendMode(DX_BLENDMODE_ALPHA, 255 * (1 - progress / LIFETIME));
+    SetDrawBlendMode(DX_BLENDMODE_ALPHA, 255 * (int)(1 - progress / LIFETIME));
     DrawBillboard3D(BasePosition, 0.5f, 0.5f, size, progress, explosionHandle, TRUE);
     DrawSubExpl(Position1, Direction1);
     DrawSubExpl(Position2, Direction2);
@@ -19,7 +19,7 @@ bool explosionEffect::DrawSingleExplosion()
     progress += oneFlame * timeScale;
     size += (20 * oneFlame) * timeScale;
     SetWriteZBuffer3D(FALSE);
-    SetDrawBlendMode(DX_BLENDMODE_ALPHA, 255 * (1 - progress / (LIFETIME-1.5f)));
+    SetDrawBlendMode(DX_BLENDMODE_ALPHA, 255 * (int)(1 - progress / (LIFETIME-1.5f)));
     DrawBillboard3D(BasePosition, 0.5f, 0.5f, size, progress, explosionHandle, TRUE);
     SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
     SetWriteZBuffer3D(TRUE);
